+++ lib/configurate/provider/dynamic.rb
@@ -7,17 +7,17 @@ module Configurate; module Provider
     def initialize
       @settings = {}
     end
-    
-    def lookup_path(setting_path, *args)      
-      key = setting_path.to_s
-      
+
+    def lookup_path(setting_path, *args)
       if setting_path.is_setter? && args.length > 0
         value = args.first
         value = value.get if value.respond_to?(:_proxy?) && value._proxy?
-        @settings[key] = value
+        *root, key = setting_path.to_a
+        hash = root.inject(@settings) {|hash, key| hash[key] ||= {} }
+        hash[key] = value
       end
-      
-      @settings[key]
+
+      Provider.lookup_in_hash setting_path, @settings
     end
   end
 end; end